split a webapp module for jetty-ee9-websocket-jakarta-client
