+++ jetty-ee8/jetty-ee8-bom/pom.xml
@@ -112,6 +112,11 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.jetty.ee8.websocket</groupId>
+        <artifactId>jetty-ee8-websocket-javax-client-webapp</artifactId>
+        <version>12.0.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.jetty.ee8.websocket</groupId>
         <artifactId>jetty-ee8-websocket-javax-common</artifactId>
         <version>12.0.0-SNAPSHOT</version>
       </dependency>
