+++ jetty-ee8/pom.xml
@@ -333,6 +333,11 @@
         </dependency>
         <dependency>
           <groupId>org.eclipse.jetty.ee8.websocket</groupId>
+          <artifactId>jetty-ee8-websocket-javax-client-webapp</artifactId>
+          <version>${project.version}</version>
+        </dependency>
+        <dependency>
+          <groupId>org.eclipse.jetty.ee8.websocket</groupId>
           <artifactId>jetty-ee8-websocket-servlet</artifactId>
           <version>${project.version}</version>
         </dependency>
