+++ jetty-ee9/jetty-ee9-websocket/jetty-ee9-websocket-jakarta-client-webapp/src/main/java/module-info.java
@@ -11,14 +11,17 @@
 // ========================================================================
 //
 
-module org.eclipse.jetty.ee9.websocket.jakarta.client
+import org.eclipse.jetty.ee9.websocket.jakarta.client.webapp.JakartaWebSocketShutdownContainer;
+
+module org.eclipse.jetty.ee9.websocket.jakarta.client.webapp
 {
     requires org.slf4j;
 
-    requires transitive org.eclipse.jetty.ee9.websocket.jakarta.common;
+    requires transitive org.eclipse.jetty.ee9.websocket.jakarta.client;
+    requires transitive jetty.servlet.api;
 
-    exports org.eclipse.jetty.ee9.websocket.jakarta.client;
+    exports org.eclipse.jetty.ee9.websocket.jakarta.client.webapp;
 
-    provides jakarta.websocket.ContainerProvider with
-        org.eclipse.jetty.ee9.websocket.jakarta.client.JakartaWebSocketClientContainerProvider;
+    provides jakarta.servlet.ServletContainerInitializer with
+        JakartaWebSocketShutdownContainer;
 }
