+++ jetty-ee9/jetty-ee9-websocket/jetty-ee9-websocket-jakarta-client-webapp/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,0 +1,1 @@
+org.eclipse.jetty.ee9.websocket.jakarta.client.webapp.JakartaWebSocketShutdownContainer
