+++ jetty-ee9/jetty-ee9-websocket/jetty-ee9-websocket-jakarta-client/pom.xml
@@ -35,7 +35,7 @@
     <dependency>
       <groupId>org.eclipse.jetty.toolchain</groupId>
       <artifactId>jetty-jakarta-servlet-api</artifactId>
-      <optional>true</optional>
+      <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty</groupId>
@@ -70,8 +70,7 @@
               osgi.extender; filter:="(osgi.extender=osgi.serviceloader.registrar)";resolution:=optional
             </Require-Capability>
             <Provide-Capability>
-              osgi.serviceloader;osgi.serviceloader=jakarta.websocket.ContainerProvider,
-              osgi.serviceloader;osgi.serviceloader=jakarta.servlet.ServletContainerInitializer
+              osgi.serviceloader;osgi.serviceloader=jakarta.websocket.ContainerProvider
             </Provide-Capability>
           </instructions>
         </configuration>
