--- jetty-ee9/jetty-ee9-websocket/jetty-ee9-websocket-jakarta-client/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,0 @@
-org.eclipse.jetty.ee9.websocket.jakarta.client.JakartaWebSocketShutdownContainer
