+++ jetty-ee9/jetty-ee9-websocket/jetty-ee9-websocket-jakarta-tests/pom.xml
@@ -24,7 +24,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.ee9.websocket</groupId>
-      <artifactId>jetty-ee9-websocket-jakarta-client</artifactId>
+      <artifactId>jetty-ee9-websocket-jakarta-client-webapp</artifactId>
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.ee9.websocket</groupId>
